refactor(core): dedupe identical claude + cursor parsers
claude and cursor stream-json parsers were functionally identical
(198 lines each). extract a shared parser at
packages/core/src/providers/shared/anthropic-envelope-parser.ts and
reduce both to thin wrappers that call it with adapter-specific
options (adapter id + log tag).

public exports (parseStreamJsonLine, parseCursorStreamLine,
ParseContext) unchanged. tests pass without modification.

396 -> 225 lines.

Closes #401
